Snap with respect to the window spacing when resizing windows.
When resizing, snap to windows and struts by taking the new window
spacing option into account.

When snapping to a window, snap with a distance of window spacing to
the other window's edge if we're approaching this edge from outside
(e.g. window A is placed left of window B, window A is enlarged and the
right edge of window A snaps to the left edge of window B minus the
spacing). If we're approaching a window edge from inside, snap to this
edge without spacing (e.g. window A is placed inside window B, window A
is enlarged an the right edge of window A snaps to the right edge of
window B).

When snapping to struts, allow both snapping directly to the edges and
snapping with spacing.