Add special icons for the button layout in settings.
This makes it a lot easier to distinct the various actions.
Icons based on the default theme icons, created by Simon Steinbeiß.

Also add the active and hidden layout in a sizegroup and drop
the fixed size, this makes the buttons a bit bigger, but this
way the icons are not chopped.

Also rename the Hide action to Minimize, this is how it's call in
the menu too.