Select window button on mouse over
When the mouse goes over a window button in the tabwin, give that
window button the focus and select it. This way the mouse functions
the same way as the keyboard. When the user clicks on a window
button, focus and raise that winodw and get rid of the tabwin.

Signed-off-by: Simon Steinbeiss <simon.steinbeiss@elfenbeinturm.at>