Tabwin list mode relative label width
The width of the application label in the tabwin window while in
list mode is now relative to the number of columns displays. This
is scaled to the width of the smallest monitor attached. Now there's
space for more label info when there are fewer columns and it properly
displays all the windows when there are lots of them open.

Signed-off-by: Simon Steinbeiss <simon.steinbeiss@elfenbeinturm.at>