Determine the maximum host name length correctly.
The actual limit on Linux is 64 characters, and on BSD it is 255
characters. The limit is defined in HOST_NAME_MAX on Posic operating
systems, so use that definition instead of the one hardcoded into
Xfwm, and print a warning if it still fails.

Too long hostnames would previously cause the hostname to be set
to null, which would cause a segfault when terminating a client.
This patch also adds a null check to that function.

Signed-off-by: Simon Steinbeiss <simon.steinbeiss@elfenbeinturm.at>