Extremely paranoid hostname lookup (bug #11289)
Get the hostname into a large temporary array first, then g_strdup it to
avoid memory wastage. The temporary array is allocated with g_new0 to ensure
it doesn't contain garbage if gethostname() does nothing. The last element
is always set to null in case gethostname() overruns the buffer and doesn't
return an error.

Signed-off-by: Simon Steinbeiss <simon.steinbeiss@elfenbeinturm.at>