Extended XSync counter
Add partial support for extended XSync counter as described
by Owen here:

http://fishsoup.net/misc/wm-spec-synchronization.html

Unfortunately, xfwm4's XSync implementation does not work too
well with some apps (e.g. gnome-terminal) when the compositor
is disabled, some apps do not seem to update the counter as
expected and the alarm is not triggered, causing a timeout.

I am not even sure supporting the extended XSync counter makes
much sense without a compositor so it's not too much of a trouble.

To limit the effect of the freeze until the timeout fires up,
reduce the limit to 500 ms.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>