Fix cycle vs. focus follow mouse
Bug: 11166

Enabling selection with the mouse in the cycle/wab window implies
that events are routed to their usual,  respective windows, meaning
that enter/leave notify events also reach regular windows.

An unfortunate side effect is that, in focus follow mouse, focus
could be sent back to the window underneath the pointer instead
of the desired selected window if the pointer was located inside
the cycle/tab window when it got unmapped.

To avoid this, track enter/leave notify events in the cycle/tab
windows and swallow the enter notify event resulting from the
cycle/tab window being unmapped.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>