Send debug messages to a log file
Bug: 10564

Logging to an X terminal while the X servder is grabbed (which happens
quite often in a window manager) may lead to a dead lock.

Only way to reliably avoid this is to send log and debug messages to a
file instead.

Log file can be specified using the environment variable XFWM4_LOG_FILE.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>