Search harder for screen for a window
Bug: 11664

_NET_REQUEST_FRAME_EXTENTS is a request sent before the window is mapped
(it's the actual purpose, estimate what the frame extents will be once
it's mapped), but as the window is not yet mapped, it is unknown to the
window manager.

That would cause _NET_REQUEST_FRAME_EXTENTS to be ignored, and some
apps/tookit (namely GLFW) rely on it and would block indefinitely until
the reply from the window manager is received.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>