Fix gravity
Bug: 11693

Fix a long standing bug with gravity which was partly due to a wrong
interpretation of the gravity definition from bug 3634 where the
reproducer provided was actually wrong.

Gravity does not change the actual reference point of the client
window, but rather "specifies how and whether the client window wants
to be shifted to make room for the window manager frame" (from the
ICCCM definition).

This makes xfwm4 behave the same as other window managers such as
metacity/mutter or kwin with regard to window gravity.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>