Check class hint name for same applications
Bug: 11685

There was some confusion in the definition of "same application" when
cycling through windows of the same application.

Same application meant windows which share a common leader or from the
same group, which basically means the same running application, not
two different instances of the same program.

Widen the search criteria to the WM class "resource name" which
contains the application name so that multiple different  instances of
the same program/application can be included.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>