focus: Ignore zero timestamp from s/n
Applications may set their _NET_WM_USER_TIME to 0 to indicate that the
window should not initially focused when it is mapped, but there is no
indication that startup-notification given timstamp should follow the
same rule.

Ignore zero timestamp from startup-notification, it will help with apps
not being focused when started from the panel or the desktop.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>