Only check for known buttons in _NET_WM_MOVERESIZE
For _NET_WM_MOVERESIZE requests, if the given button does not match any
available physical button, we would wait for that (impossible) button
combination to be released.

Check that the given button is one of the well known button or just wait
for any button release otherwise.

That fixes an issue with KDE apps where clicking on the menu bar to move
the window would hang forever after the mouse button is depressed.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>