netwm: reset decoration on netwm type change
Bug: 10413

Some applications may temporarily set a transient relationship betwen
their toplevel windows to maintain a stacking order, in which case we
would change the decorations (as transients may not have all buttons
available).

However, we ought to restore the controls when this transient
relationship is removed by the application, otherwise we leave the
toplevels with reduced controls.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>