stacking: Raise on focus to activate
Bug: 12634

xfwm4 would raise a client upon receiving the _NET_ACTIVE_WINDOW message
but if the client provided an old timestamp, the window won't be focused
(as the timestamp will be older than the last current X server time).

To avoid this, only raise the window on focus event notification and
when using click-to-focus mode.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>