client: Maximize on expected output
Bug: 13284

When dragging a window from a large monitor to the top of a smaller monitor to
maximize the window, it is possible that the window gets maximised on a
different monitor then the monitor your mouse is.

This happens because clientToggleMaximized uses the center point of the window
to calculate on which monitor the window should be maximised.

To fix this behaviour this commits adds a clientToggleMaximizedAtPoint method
so we can tell on which monitor the window should be maximised.