moveresize: Cancel move/resize if client is gone
Bug: 13344

If a window is unmapped while the user is moving or resizing it, we
would cancel the move but swallow the event, which would leave the
window frame on screen as the main event handle would not get the
unmap notify.

Mark the client as gone when detected in the move/resize event loop,
let the event get through to the main event handler that will free the
client and shortcut the move/resize operation if the client is marked
as gone.

(cherry picked from commit 5155c86cfc198946fdf9d55f88b187a58b0340fc)