screen: Fix broken monitor index
Bug: 13608

Commit 5895577 (Use ximerama indices for _NET_WM_FULLSCREEN_MONITORS)
introduced a regression because the monitor index was wrongly compared
against the total number of monitors, which would falsely assume that
the Xrandr query failed whereas it actually worked.

Fix the false comparison and also use the screen size as fallback values
in case the Xrandr really failed so that we don't end up using random
uninitialized values.