compositor: Disable XPresent on error
Bug: 13607

If XPresentPixmap() fails, we would disable Present rather that waiting
for a notification that will never arrive.

But that relies on the actual window to be reported in the XError, which
is not necessarily the case.

Disable Xpresent for all screens as we are not sure we can tell which
one from the resource id.