settings-dialog: Fix drag and drop of buttons
Bug: 13861

Drag and Drop inside GtkSocket and GtkPlug doesn't work correctly.

The drag_end handler isn't called when a button from "Button layout" is
dragged.

Fortunately, this only happens when the button is dragged onto the
specified drag area.

All we need to do is move the gtk_widget_show() function at the end of
the drag-data-received handler.