screen: Use computed screen size instead of xlib
When the xrandr layout is changed, we recompute the screen size on the
actual monitors size and layout.

Use this computed value instead of the size reported by Xlib as it's not
yet updated by the time we reconfigure our display. This is particularly
bad when the compositor is enabled as the overlay window end up being of
the wrong size.

(cherry picked from commit 480517d49d3e4cf48357fbe47d193f62542ab94b)