stacking: Raise ancestor on activate if different from window
Bug 14164

This is a regression introduced by 3877bf45, where we would differ
raising the window until it's focused.

When using click to focus, activating a transient window would raise
only that transient and not its ancestor, if any.

Change the logic to raise the ancestor if it's different from the
requested window itself.