Add back the glXWaitGL to compositor sync code
In d6e7fbc4 the glXWaitX/glXWaitGL synchronization around compositor
texture upload was replaced by an XSync fence. This causes extreme
flickering on nvidia. My current understanding is that the fence code
is equivalent to glXWaitX, so the other half of the synchronization
is not handled. This patch restores glXWaitGL to deal with that.

Note that if my understanding is correct then this patch will
behave exactly the same as reverting d6e7fbc4. This means the
problems d6e7fbc4 intended to fix will come back.

Note (bis), I do not see any regression with intel/modesettings so I
reckon it's fine to land -- (ofourdan)