events: ignore client configure requests when maxized
Bug: 13954

Some clients tend to send us a configure request when transitioning to
maximized, which introduces a race condition because the window manager
will remove the maximized flag when the client resizes itself, now that
we allow for that since commit aee8b25a.

Ignore the configure request from the client when maximized or
fullscreen to work around such behavior from the client.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>