stacking: Raise all transients together
Bug 15303

Xfwm4 would raise the transients of a given window along with the parent
window, to keep the transients above their parents, but would not raise
the parents along with the transient windows.

Rework the stacking code to raise the deepest parent of a transient
being raised so that all parents window of a transient get raised along
with it.