moveresize: Size increment depends on scale
GTK would update the size increment to match the scale, so when using a
scale 2x the size increment of all windows is set to 2 as well.

As we use the size increment to decide whether to show the position and
resize indicator window, we would end up showing it for all windows
when scale is 2.

Take the scale into account for size increment as well.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>