compositor: Do not read from front buffer
Reading from the front buffer is not advisable and gives random results
depending on the driver and/or hardware.

So instead of relying on the front buffer, re-copy the entire content to
the back buffer again just after we swapped buffers.

Also, by doing so, we don't need to accumulate damages with GLX backend.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>
Fixes: a818cc0 "compositor: Optimize GLX code path"