Protect against XErrors
Places where the window manager uses the client window are all possible
sources of XError because the window may have vanished and the window
manager won't know until it gets the destroy notification from the
Xserver.

As GDK error handler is smart and uses the serial of the requests to
check for errors, if the XError is not trapped in time it will end up in
GDK and cause an abort.

Protect against such a possibility by making sure cases where the XID
of the client or one of its dependent resources is used remains withing
an error trap/push.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>