compositor: Apply texture filter for zooming
Bug: 15392

The texture filter changes based on the zoom value to give best results
when zooming at higher levels.

As we do not unbind/rebind at each repaint anymore, the texture filter
would not be updated, hence the chosen filter is ignored.

Make sure we apply the texture filter when zooming.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>