compositor: Remove glXWaitGL() call
Bug: 15453

This was added to avoid flickering with NVIDIA proprietary/closed source
driver (commit f75652e - "Add back the glXWaitGL to compositor sync
code"), but that translates as a busy wait, hence increasing CPU usage
of xfwm4 with the NVIDIA proprietary/closed source driver.

By setting “__GL_MaxFramesAllowed” to 1, that busy wait should not be
needed anymore.

As a bonus, it also improve responsiveness with with the open source
drivers.

Taken from similar patch posted by NVIDIA developer for kwin:
https://phabricator.kde.org/D19867

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>