compositor: Fix GL texture filter
Bug: 15676

Between versions 4.13.1 and 4.13.2 the GL_NEAREST filter (used on the
deepest levels of desktop zoom) stopped working, and GL_LINEAR was
used instead. Fix by setting GL_TEXTURE_MIN_FILTER in addition to
GL_TEXTURE_MAG_FILTER, like was done in version 4.13.1; apparently the
former needs to be correctly configured for the latter to take effect.

Signed-off-by: Juha Aatrokoski <jha@iki.fi>