common: There might be no primary monitor
Bug: 15852

The function `xfwm_get_primary_monitor_geometry()` assumes there is
always a primary monitor, which is not true if no monitor is connected,
in which case the size is not meaningful.

Make `xfwm_get_primary_monitor_geometry()` return a boolean value that
can be used by the caller to determine if the reported size is
meaningless.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>
(cherry picked from commit 699c0aeb782cb36181d2cd6fe39804363c23fb8c)