compositor: Use actual surface position and size
Bug: 15966

When using wireframe move/resize along with transparent window
decorations, the window would appear to move along with the frame.

This is because the move/resize operation updates the internal data
that the compositor uses for drawing the client decorations with a
different opacity.

To avoid the issue, just use the actual window positio nand size.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>