device: Use standard grabs for passive button grabs
Bug: 16347

XInput2 passive button grab would not fail if the button is already
grabbed by the client.

X-AIR-Edit is such a client which issues a button grab on its own
window, meaning that with XI2 button it would not react to button events
anymore.

Revert to the standard grab button for the passive grabs so that such
apps can still work with xfwm4.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>