device: Check standard passive button grabs
Bug: 16649

With Xi2 we need to use Xi2 routines.

But some client may issue standard X11 core protocol grabs the on the
client window.

Grabs from the standard core protocol and Xi2 grabs do not mix well.

So instead of ditching the Xi2 grabs, we issue the core protocol first,
check that it succeeds and only if that works we can issue the Xi2
grabs.

That keeps the Xi2 grabs logic while fixing the issue with X-AIR-Edit
which uses a passive grab on the pointer using core X11 protocol.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>