transients: Do no search for parent in lower layers
Bug: 15891

When raising a transient window we would automatically raise its parent
windows.

However, if the transient and parent are not placed on the same layer,
we might raise the parent from a lower layer.

To avoid that issue, limit the scope of search to the same or higher
layers.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>
(cherry picked from commit 3d6e0472e9a5ec310ff11294ae2c07dfd4a1f538)