cycle: Do not grab the pointer for alt-tab
Bug: 16708

With Xi2, the clients needs to reset the valuators (for mouse wheel)
themselves, usually on focus events.

Qt instead does that on EnterNotify, but ignores the EnterNotify events
resulting from grabs first. This is a bug in Qt as that prevents the
valuators to be reset if the pointer enters the window while grabbed.

Precisely, xfwm4 would issue a pointer grab while cycling to prevent
from random button events, and would only ungrab after the client window
is raised and focused.

While this is a bug in Qt, we can still work around the issue in xfwm4
by not issuing a pointer grab and simply abort the window cycling when a
click occurs outside of the alt-tab window.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>

(cherry picked from commit d2cf2e80ee0c997ae951282a611fadbd8c832094)