main: Don't use the compositor if not available
Bug: 16713

If one of the required X11 extension is not available, the compositor
cannot be used.

Yet xfwm4 would still try to enable the compositor for each screen,
which would not work and prevent xfwm4 from starting.

If we can't use the compositor, well, we'd better not use it...

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>

(cherry picked from commit 2c78273d35abc60af697b571713809b00b40ce38)