client: Rely on XRes reported PID if available
When xfwm4 needs to kill an unresponsive client, it will use the its PID
is known.

However, the _NET_WM_PID as defined in the extended window manager hint
mechanism is unreliable when the client runs within a sandbox, as it is
the client itself who sets that property using its own view of the PID.

The Xserver can provide a more accurate value of the client PID using
the XRes X11 extension, so use that protocol if available.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>
(cherry picked from commit 49c1a33a4dfe15f312334584c22b4ec3ed792214)