client: Really send synthetic configure notify
Commit 821ee7cc would not actually send the synthetic configure notify
events anymore. That obviously breaks plenty of applications…

While the logic of the commit itself is sound, the problem is that it
bases the descision to send the configure notify event by comparing
against the current window size and location.

The problem is that the current window size and location is modified in
various places, defeating the logic.

Add an “applied_geometry” rectangle that we update only in a single
place in the code, so that we know exactly when to send a synthetic
configure notify event.

Closes: https://gitlab.xfce.org/xfce/xfwm4/-/issues/403
Signed-off-by: Olivier Fourdan <fourdan@xfce.org>