Fix errorTrap leak in free_win_data
The function free_win_data() may generate X11 errors as the client
window may already be gone.

To avoid those, the entire function code is supposed to be enclosed
within an error trap/push.

A typo in the code would however push twice instead of a balanced
push/pop which causes the list of event sequences in GDK to grow
indefinitely, leading to slowdowns as the list grows over time.

Fix the leak by balancing the error push with an error pop at the end
of the function.

https://gitlab.xfce.org/xfce/xfwm4/-/issues/351