Add define for compositor_timeout_cb interval
The call to g_timeout_add in add_repair seems to be mixing up priority
and callback interval.  Instead, let's call g_timeout_add_full to make
it clear there are two separate fields, and add an appropriate define
for the callback interval, setting it to the same 1ms value it was
currently using.

Signed-off-by: Michael Banack <bob5972@banack.net>