compositor: Make cursor zoom optional
Xfwm4 4.14 made the cursor zoom along with the desktop. While this
sounds like a sensible thing to do, it's not necessarily an improvement
for accessibility, because an enlarged cursor also hides what's being
zoomed underneath.

Add a new option to control whether the pointer is zoomed along with the
rest of the desktop.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>
Closes: https://gitlab.xfce.org/xfce/xfwm4/-/issues/439