compositor: Add support for opaque regions
Opaque region is a hint set by the clients via the _NET_WM_OPAQUE_REGION
property to indicate the compositor which regions of an ARGB window are
considered opaque and therefore can be culled out during repaint.

While this is of little use with regular RGB windows, with the planned
move to client-side decorations in xfce, we'll have more and more ARGB
windows, making such an optimization a beneficial improvement.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>