compositor: Damage opaque region if changed
When using multi buffers (Present, GLX), if the client updates its
opaque region in between (e.g. when resizing the window) we may cull out
regions on screen that won't get updated, leaving trails of unpainted
areas.

To avoid that, make sure we add damage for the delta in opaque region.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>