client: Send synthetic configure on force redraw
xfwm4 sends a synthetic configure event when required as per the ICCCM
specification, but a force redraw as when maximizing a window while
moving generates a fake move, which defats the logic.

Make sure to send a synthetic configure event on force redraw, that
fixes an issue with VTE terminals (like xfce4-terminal) which do not
resize correctly if maximized while pushing toward the top of the
screen.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>