compositor: Damage on opaque region update only if visible
If a client updates the opaque region of its window, regions previously
opaque may become transparent, meaning that they would need to be
repainted to let other windows underneath see through.

The compositor already computes that and adds damage for the relevant
regions.

Make sure to do that only for visible and redirected windows, other
windows shouldn't trigger that though.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>