compositor: Limit opaque region clipping to window extents
The opaque region is set and updated by the client, which may lag when
the window is resized.

That may leave unpainted areas when resizing a window as the opaque
region set by the client could be actually larger than the window itself
and hence prevent areas to be updated while they should, causing
artifacts on screen.

Make sure to limit the clipping of the opaque region by the window
extents to avoid the issue.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>