Revert "compositor: Do not damage on opaque region update"
Thunderbird is acting really weird with its opaque region, it sometimes
set a (wrong) opaque region for a very short time during resize and
immediately remove it, but that can prevent non-opaque regions such as
drop shadows with client side decorations from being damaged correctly.

While this is clearly a client bug, it shows that damaging the
difference on opaque region update can be useful.

This reverts commit 1e80481ac8da9cdeddc204c2c8e9c2f5841564ef.