netwm: Mark splashscreen as transient for groups
KDE applications such as Krita place their splashscreen window above
other windows, but may display a dialog at startup, and that dialog
would end up underneath the splashscreen.

Make the splashscreen windows transient for groups, so that if another
window from the same applications is eventually mapped, the splashscreen
will be sent back automatically at the same layer as the application and
will not hide it.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>
Fixes: https://gitlab.xfce.org/xfce/xfwm4/-/issues/470