events: Allow resizing regardless of modifiers
xfwm4 expects no keyboard modifiers (other than the usual locks) in
button events to allow interactive resizing.

There is no reason to be so picky, especially considering there is no
such check when moving windows.

Relax the requirements for resizing and allow interactive resize with
the mouse regardless of the modifiers being pressed.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>
Closes: https://gitlab.xfce.org/xfce/xfwm4/-/issues/487