netwm: Do not use _NET_WM_PING in multi-screen
GDK dropped support for multi-screen several years ago.

Unfortunately, the code in GDK's _gdk_wm_protocols_filter() checks for
the default root window when dealing with _NET_WM_PING messages, and
that will cause an event loop when trying to use it on a multi-screen
setup.

Avoid the issue by not using _NET_WM_PING on multi-screen setups.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>
Closes:https://gitlab.xfce.org/xfce/xfwm4/-/issues/506