display: check for XResQueryClientIds to enable XRes extension.
The use of XRes extention depends libXRes-devel, but in xfwm4 looks for the
XResQueryClients symbol to enable its use, however this one exists in libxres
1.0.7, but xfwm4 uses other symbols that were only introduced in 1.2, failing
to compile xfwm4 with this version.

I change the search for the symbol XResQueryClients to XResQueryClientIds also
introduced in 1.2, being able to compile xfwm4 with two versions of libXRes.

Fix issue #505

Signed-off-by: Matias De lellis <mati86dl@gmail.com>