compositor: Disable GLX on mutli-screen (Zaphod)
Zaphod mode support was removed in GTK3 a while ago, support for
multiple-screen in xfwm4 is only on a best effort basis.

GLX doesn't seem to work reliably when using multiple X11 screens (aka
Zaphod) and produces a black image on all but one screen.

Disable GLX on the additional X11 screens.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>