compositor: Do not set deprecated compositor selection
Zaphod mode support was removed in GTK3 a while ago, support for
multiple-screen in xfwm4 is only on a best effort basis.

The deprecated compositor atom selection used to inform that a
compositor is running opn a display is not mutli-screen (aka Zaphod)
compatible.

As a result, xfwm4 would falsely detect a compositor already running
when dealing with multiple X11 screen wheres it was its own selection.

To avoid the issue, avoid setting the deprecated atom selection and use
only the newer one which is mutli-scren aware.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>