placement: Ignore struts from windows on other monitors
The EWMH[1] “specification” defines truts are relative to the entire X11
screen, which include all monitors.

Under some configuration, when montors are placed one on top of the
other, that struts set on the monitor above would invalidate the entire
lower monitor, preventing from placing windows on that monitors.

To avoid that issue, xfwm4 was computing the actual struts area and
would simply ignore any struts greater than a quarter of the screen.

This is largely inefficient and only very partially solve the problem.

Instead, simply apply struts for windows on that monitor and ignore
those from the other monitors, so that a panel placed on the monitors
above would simply be ignored on the monitor below.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>
Closes: https://gitlab.xfce.org/xfce/xfwm4/-/issues/517