placement: Restack fullscreen window when necessary
When a window is fullscreen, it's placed on a layer above all the
others.

When a fullscreen window loses focus, however, xfwm4 will send that
window back to its original layer so that the other windows which may
have focus can be seen and not hidden underneath the fullscreen window.

When using focus follow mouse mode without raise window on focus, the
actual windows stacking is not changed however on focus change, meaning
that transferring focus to another window may cause a lot of windows
above the fullscreen window which just lost focus.

To avoid that, only restack the fullscreen window when the newly focus
window overlaps with it. This way, focusing a window on another monitor
will not make other windows come above the fullscreen window.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>
Closes: https://gitlab.xfce.org/xfce/xfwm4/-/issues/601