client: Do not hide transients for group if above
Firefox sets the type "utility" for its Picture-in-Picture window but
does not make it transient to any window.

xfwm4 would treat this as a transient for group and therefore minimize
that utility window along with the rest of the Firefox windows, meaning
that minimizing Firefox would also hide the Picture-in-Picture window.

That's actually a desirable behavior for most clients (e.g. The GIMP)
but it looks Firefox got tweaked to rely on a specific implementation,
meaning that people actually expect the same behavior with xfce.

Simply ignore the utility windows if placed on a different layer.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>
Closes: https://gitlab.xfce.org/xfce/xfwm4/-/issues/525