compositor: Damage window when adding the opaque region
When mapping a new window for the first time, the client may set an
opaque region prior to draw anything.

As opaquere regions are meant to hint the compositor that anything
underneath doesn't need updating (as hidden behind an opaque region,
precisely), xfwm4 would leave that region untouched.

However, when using double buffering as with xpresent, it means that
content of the different buffers may end up being different if there is
any damage occurring underneath the opaque region, before the client
itself has painted anything.

As the buffers alternate quickly, the content behind the opaque region
will show alternatively which is ugly (although technically correct).

To avoid that issue, when setting the opaque region first, make sure to
damage the whole window so that the content of the different buffers
remain updated even if behind the opaque region.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>
Closes: https://gitlab.xfce.org/xfce/xfwm4/-/issues/583
Fixes: 50d34206 - Damage opaque region if changed