focus: Fix focus being lost on workspace switch
When switching workspaces using the keyboard, the focus is lost and
any later keyboard shortcuts fail.

This is because of commit 3ff334f50 - "Check for user focus for when
passing focus".

The problem is that the code checks for the current focused (or pending)
window, but does not actually verify that neither are non NULL.

This partially reverts commit 3ff334f50

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>