workspace: Update maximized on struts change
Since commit 7e8d60bb (Ignore struts from windows on other monitors),
the work area is limited to the primary monitor ony.

As a result, when a panel is placed on a non-primary output, the workarea
does not actually change.

But we would update the maximized window size and position only when the
workarea changed, which is now wrong, we need to update any time the
struts change.

Closes: https://gitlab.xfce.org/xfce/xfwm4/-/issues/658
Fixes: 7e8d60bb
Signed-off-by: Olivier Fourdan <fourdan@xfce.org>