introduce window fences
This new feature allows defining screen regions that act like fences around
grassland, so cattle remains confined in there: when a window is moved over
some fenced area, it will snap into it and expand to the fenced region.

The window later can be moved out to an unfenced area, where it reverts to
normal behaviour, or moved into another one.

Configuration is done via xfconf in the xfwm4 channel:

  |
  +-- "fences" [empty/array]
  |   |
  |   +-- <ident1> [array]
  |   |   |
  |   |   +-- "title" [string]
  |   |   +-- "geometry"         [string]  -- X geometry string (see XParseGeometry)
  |   |   +-- "border-color"     [string]  -- Gdk color string for border
  |   |   +-- "background-color" [string]  -- Gdk color string for background
  |   |   +-- "auto-maximize"    [boolean] -- auto-maximize on entering fence
  |   +-- <ident> [array]
  |   |   | ...

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>