netwm: record window type name in client struct
For now we're just recording the XAtom ID of the window type wmhint,
which is used for window type dependent actions (eg. keeping dialogs
on top). But when the actual string is required, it's getting a bit
complicated (have to scan through the whole array). Thus record the
name right when the wmhint is updated and we have to compare anyways.

X-Feature-Add: record-window-type-name
X-Base: master
Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>